Repaint class cleanup, found memory leak
Stacktrace refactored
