+++ CMakeLists.txt
@@ -181,6 +181,7 @@ if (BUILD_POPPLER)
 # Comment out, seems to be a problem on Debian...
 # Do we need to detect something to enable / disable?
 #    -lopenjpeg
+    -lbacktrace
     ${JPEG_LIBRARIES}
     ${PopplerLibs_LDFLAGS}
     ${lcms_LDFLAGS}
