+++ src/Xournalpp.cpp
@@ -19,11 +19,6 @@ int main(int argc, char* argv[])
 {
 	// init crash handler
 	installCrashHandlers();
-	if (argc)
-	{
-		// Filename is needed to get backtracke with filenumbers
-		Stacktrace::setExename(argv[0]);
-	}
 
 #ifdef DEV_CALL_LOG
 	Log::initlog();
