+++ src/util/Stacktrace.h
@@ -11,8 +11,6 @@
 
 #pragma once
 
-#include "StringUtils.h"
-
 #include <iostream>
 
 class Stacktrace
@@ -22,8 +20,10 @@ private:
 	virtual ~Stacktrace();
 
 public:
-	static void setExename(string name);
-
 	static void printStracktrace();
 	static void printStracktrace(std::ostream& stream);
+
+private:
+	static void errorCallback(std::ostream* stream, const char* msg, int errnum);
+	static int fullCallback(std::ostream* stream, uintptr_t pc, const char* filename, int lineno, const char* function);
 };
