+++ src/util/XournalTypeList.h
@@ -226,5 +226,6 @@ XOJ_DECLARE_TYPE(ImageElementView, 215);
 XOJ_DECLARE_TYPE(PdfElementView, 216);
 XOJ_DECLARE_TYPE(MainWindowToolbarMenu, 216);
 XOJ_DECLARE_TYPE(LastSelectedTool, 217);
+XOJ_DECLARE_TYPE(RepaintWidgetHandler, 218);
 
 //XOJ_DECLARE_TYPE(XXXXXXXXXXXXXXXX, 206);
